+++ sui-java-service/suiDemoContracts-cmd-rest/src/main/java/org/dddml/suidemocontracts/sui/contract/order/OrderItemRemoved.java
@@ -17,7 +17,7 @@ import java.util.*;
 public class OrderItemRemoved {
     private String id;
 
-    private Long version;
+    private BigInteger version;
 
     private String productId;
 
@@ -29,11 +29,11 @@ public class OrderItemRemoved {
         this.id = id;
     }
 
-    public Long getVersion() {
+    public BigInteger getVersion() {
         return version;
     }
 
-    public void setVersion(Long version) {
+    public void setVersion(BigInteger version) {
         this.version = version;
     }
